added PCA_python.py
PCA(Principal Component Analysis):-

Principal Component Analysis (PCA) is a
dimension-reduction tool that can be used to
reduce a large set of variables to a small set
that still contains most of the information in
the large set.
• Principal component analysis (PCA) is a mathematical
procedure that transforms a number of (possibly)
correlated variables into a (smaller) number of
uncorrelated variables called principal components.
• The first principal component accounts for as much of
the variability in the data as possible, and each
succeeding component accounts for as much of the
remaining variability as possible.
• Principal components analysis is similar to another
multivariate procedure called Factor Analysis. They are
often confused and many scientists do not understand
the difference between the two methods or what types
of analyses they are each best suited.