Create KaprekarNumber.java
To check if an n-digit number k is a Kaprekar number or not. Square it and add the right n digits to the left n or n-1 digits. If the resultant sum is k, then k is called a Kaprekar number. For example, 9 is a Kaprekar number since 9^2 = 81 & 8 + 1 = 9;  297 is a Kaprekar number as 297^2 = 88209 & 88 + 209 = 297.