Create InversionCount.cpp
Counting inversions Inversion Count for an array indicates – how far (or close) the array is from being sorted. If the array is already sorted then the inversion count is 0. If the array is sorted in the reverse order that inversion count is the maximum. Formally speaking, two elements a[i] and a[j] form an inversion if a[i] > a[j] and i < j.
Example: 
Input: arr[] = {8, 4, 2, 1} 
Output: 6. The six inversions: (8,4), (4,2), (8,2), (8,1), (4,1), (2,1). 
Input: arr[] = {3, 1, 2} 
Output: 2. The two inversions: (3,1), (3,2).