Create KEYLOGGER IN PYTHON .PY
keylogger in python .
Keystroke logging is the process of recording (logging) the keys pressed on a keyboard (usually when the user is unaware). It is also known as keylogging or keyboard capturing.
These programs are used for troubleshooting technical problems with computers and business networks. It can also be used to monitor network usages but more often than not it is used for malicious intents like stealing passwords.
This article illustrates designing a keylogger for windows and Linux.
 

Keylogger for Windows

Download some python libraries 
1) pywin32 
2) pyhook‘
Following is the code to create a keylogger in python