Create easy and efficient nqueen algorithm.py
easy and efficient n queen algorithm.
The N Queen is the problem of placing N chess queens on an N×N chessboard so that no two queens attack each other
Naive Algorithm 
Generate all possible configurations of queens on board and print a configuration that satisfies the given constraints.


Backtracking Algorithm 
The idea is to place queens one by one in different columns, starting from the leftmost column. When we place a queen in a column, we check for clashes with already placed queens. In the current column, if we find a row for which there is no clash, we mark this row and column as part of the solution. If we do not find such a row due to clashes, then we backtrack and return false.
Time Complexity: O(N!)
Auxiliary Space: O(N2)

Optimization in is_safe() function 
The idea is not to check every element in right and left diagonal, instead use the property of diagonals: 
1. The sum of i and j is constant and unique for each right diagonal, where i is the row of elements and j is the 
column of elements. 
2. The difference between i and j is constant and unique for each left diagonal, where i and j are row and column of element respectively.
Implementation of Backtracking solution(with optimization