Create Leaders_in-an_Array.py
A leader in an array is an element which is greater than all the elements on its right side in the array.

Methodology :
Run two loops. The outer loop runs from the left to the right of the array, in which i will point to the index of each element of the array. The inner loop runs from the i+1th position to the end of the array, in which j will point to the index of each element on the right side of the element picked by the outer loop. If the element picked by the outer loop is greater than all the elements on its right side picked by the inner loop, then the element is the leader.

Pseudo Code
for i=0 to n-1
    for j = i to n-1
        if array[i] < array[j]
            break
        if j == n-1
            print array[i]

Time Complexity
Since there are 2 for loops, time complexity is O(N x N).