Create MERGE.cpp
Approach:
 This problem can be solved using dynamic programming. Let’s define the states of the DP first. DP[l][r] will be the minimum cost of merging the subarray arr[l…r] into one. Given a list of N integers, the task is to merge all the elements of the list into one with the minimum possible cost. 
The rule for merging is as follows: 
Choose any two adjacent elements of the list with values say X and Y and merge them into a single element with value (X + Y) paying a total cost of (X + Y). The time complexity of the approach will be O(N3) as there are N * N states in total and each state takes O(N) time to solve in general.