Rename projectsort.py to leaders/python/leaders in a array.py
In computer programming, an array is a collection of elements of the same data type, stored in contiguous memory locations, and accessed using an index or a subscript. A leader in an array is an element that is greater than all the elements in the array.

To find the leaders in an array, we can start iterating through the array from the right end and keep track of the maximum element seen so far. If the current element is greater than the maximum element seen so far, it is a leader, and we add it to a list of leaders. We continue this process until the leader is found.